fixed set store_id as smallInteger

+++ Indexer/Database/Migrations/2018_06_12_132242_create_product_category_index_tables.php
@@ -17,7 +17,7 @@ class CreateProductCategoryIndexTables extends Migration
             $table->increments('id');
 
             $table->unsignedInteger('product_id');
-            $table->unsignedInteger('store_id');
+            $table->smallInteger('store_id');
 
             $table->timestamps();
 
@@ -29,7 +29,7 @@ class CreateProductCategoryIndexTables extends Migration
             $table->increments('id');
 
             $table->unsignedInteger('category_id');
-            $table->unsignedInteger('store_id');
+            $table->smallInteger('store_id');
 
             $table->timestamps();
